If runtime is java, set runtime config to openjdk8 in app.yaml (#161)
This is a change that we should push up to appcfg/staging or into
gcloud. No tool at the moment appears to be building a dockerfile
with runtime:java and vm:true that is correct. By forcing the jdk
to be openjdk8 we trigger the creation of a Dockerfile with the
correct java8jetty9 image.